% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneScan3D.R
\docType{data}
\name{Example.GeneScan3D}
\alias{Example.GeneScan3D}
\alias{GeneScan3D.example}
\title{Data example for GeneScan3D (gene-based testing by integrating long-range chromatin interactions).}
\format{
An object of class \code{list} of length 13.
}
\usage{
data("GeneScan3D.example")
}
\description{
This simulated example dataset contains outcome variable Y, covariate X, genotype and functional annotation matrices for gene and buffer region, promoter and two enhancers, positions of genetic variants in gene and buffer region.
}
\details{
We generated genotypes for 2,000 individuals in a 14.5 Kb gene region, a promoter as a 0.5 Kb segment upstream of the TSS (the start point of the gene) and R = 2 enhancers with length 2 KB, which are outside the 15 Kb gene plus promoter region.
}
\examples{
data("GeneScan3D.example")

Y=GeneScan3D.example$Y; X=GeneScan3D.example$X; n=length(Y)

G_gene_buffer=GeneScan3D.example$G_gene_buffer
G_promoter=GeneScan3D.example$G_promoter
G_EnhancerAll=cbind(GeneScan3D.example$G_Enhancer1,GeneScan3D.example$G_Enhancer2)

Z_gene_buffer=GeneScan3D.example$Z_gene_buffer
Z_promoter=GeneScan3D.example$Z_promoter
Z_EnhancerAll=rbind(GeneScan3D.example$Z_Enhancer1,GeneScan3D.example$Z_Enhancer2)

pos_gene_buffer=GeneScan3D.example$pos_gene_buffer
}
\keyword{data}
