% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneScan3D.R
\name{GeneScan.Null.Model}
\alias{GeneScan.Null.Model}
\title{The preliminary data management for GeneScan3DKnock.}
\usage{
GeneScan.Null.Model(Y, X = NULL, id = NULL, out_type = "C", B = 1000)
}
\arguments{
\item{Y}{The outcome variable, an n*1 matrix where n is the number of individuals.}

\item{X}{An n*d covariates matrix where d is the number of covariates.}

\item{id}{The subject id. This is used to match phenotype with genotype. The default is NULL, where the matched phenotype and genotype matrices are assumed.}

\item{out_type}{Type of outcome variable. Can be either "C" for continuous or "D" for dichotomous. The default is "C".}

\item{B}{Number of resampling replicates. The default is 1000. A larger value leads to more accurate and stable p-value calculation, but requires more computing time.}
}
\value{
It returns a list used for function GeneScan1D(), GeneScan3D() and GeneScan3D.KnockoffGeneration().
}
\description{
This function does the preliminary data management and fit the model under null hypothesis using all the covariates. The output will be used in the other GeneScan functions.
}
\examples{
library(GeneScan3DKnock)

# Load data example
data("GeneScan3D.example")
# Y: outcomes, n by 1 matrix for n=2000 individuals
# X: covariates, n by d matrix for d=1 covariate
Y=GeneScan3D.example$Y; X=GeneScan3D.example$X;

# Preliminary data management
set.seed(12345)
result.null.model=GeneScan.Null.Model(Y, X, out_type="C", B=1000)

}
