% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneScan3D.R
\docType{data}
\name{GeneScan3D.example}
\alias{GeneScan3D.example}
\title{Data example for GeneScan3D (A unified framework for gene-based testing with joint analysis of coding and regulatory variation)}
\format{
An object of class \code{list} of length 12.
}
\usage{
data("GeneScan3D.example")
}
\description{
The dataset contains outcome variable Y, covariate X, genotype data for gene buffer region, promoter and two enhancers, weight matrices for functional annotations and positions of genetic variants in gene buffer region as well as promoter.
}
\examples{
data("GeneScan3D.example")

Y=GeneScan3D.example$Y; X=GeneScan3D.example$X

G_gene_buffer=GeneScan3D.example$G_gene_buffer
G_promoter=GeneScan3D.example$G_promoter
G_Enhancer1=GeneScan3D.example$G_Enhancer1
G_Enhancer2=GeneScan3D.example$G_Enhancer2

Z_gene_buffer=GeneScan3D.example$Z_gene_buffer
Z_promoter=GeneScan3D.example$Z_promoter
Z_Enhancer1=GeneScan3D.example$Z_Enhancer1
Z_Enhancer2=GeneScan3D.example$Z_Enhancer2

pos_gene_buffer=GeneScan3D.example$pos_gene_buffer
pos_promoter=GeneScan3D.example$pos_promoter
n=length(Y)
}
\keyword{data}
