% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneScan3D.R
\docType{data}
\name{GeneScan3DKnock.example}
\alias{GeneScan3DKnock.example}
\title{Data example for GeneScan3DKnock (Integration of knockoff statistics for causal gene identification)}
\format{
An object of class \code{data.frame} with 100 rows and 7 columns.
}
\usage{
data("GeneScan3DKnock.example")
}
\description{
This example dataset contains the original and five knockoff p-values for N=100 genes. For each gene, there are gene id, original Cauchy3D p-value and five knockoff Cauchy3D p-values. The data can be used in GeneScan3DKnock() function to calculate the knockoff statistics and q-values for each gene. To generate knockoffs and obtain the knockoff p-values, please find the functions in R Package 'KnockoffScreen'.
}
\keyword{data}
