% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneScan3D.R
\name{GeneScan3DKnock}
\alias{GeneScan3DKnock}
\title{Integration of knockoff statistics for causal gene identification}
\usage{
GeneScan3DKnock(
  M = 5,
  p0 = GeneScan3DKnock.example$Cauchy3D.all.original,
  p_ko = cbind(GeneScan3DKnock.example$Cauchy3D.all.ko1,
    GeneScan3DKnock.example$Cauchy3D.all.ko2, GeneScan3DKnock.example$Cauchy3D.all.ko3,
    GeneScan3DKnock.example$Cauchy3D.all.ko4, GeneScan3DKnock.example$Cauchy3D.all.ko5),
  fdr = 0.1,
  gene_id = GeneScan3DKnock.example$gene.id
)
}
\arguments{
\item{M}{Number of multiple knockoffs. We use M=5 in our analysis.}

\item{p0}{A N-dimensional vector of the original p-values for N genes considered in the analysis. The p-values can be obtained in GeneScan3D() function or other analysis.}

\item{p_ko}{A N*M matrix of M knockoff p-values for N genes considered in the analysis. The knockoff p-values can be obtained in R package 'KnockoffScreen'.}

\item{fdr}{The false discovery rate (FDR) threshold. The default is 0.1.}

\item{gene_id}{The genes id for N genes considered in the analysis, which can also be the windows id or an indicator vector from 1 to N.}
}
\value{
\item{W}{The knockoff statistics for N genes.}

\item{Qvalue}{The Q-values for N genes.}

\item{gene_sign}{Significant genes obtained in the knockoff test with Q-values less then the fdr threshold.}
}
\description{
This function calculates the knockoff statistics and q-values after proving the original and knockoff p-values for each gene (or window).
}
\examples{
library(GeneScan3DKnock)

# Load data example
data("GeneScan3DKnock.example")

result.GeneScan3DKnock=GeneScan3DKnock(M=5,p0=GeneScan3DKnock.example$Cauchy3D.all.original,
p_ko=cbind(GeneScan3DKnock.example$Cauchy3D.all.ko1,GeneScan3DKnock.example$Cauchy3D.all.ko2,
GeneScan3DKnock.example$Cauchy3D.all.ko3,GeneScan3DKnock.example$Cauchy3D.all.ko4,
GeneScan3DKnock.example$Cauchy3D.all.ko5),fdr = 0.1,gene_id=GeneScan3DKnock.example$gene.id)

#Obtain knockoff statistics, q-values and the significant genes.
W=result.GeneScan3DKnock$W
Qvalue=result.GeneScan3DKnock$Qvalue
gene_sign=result.GeneScan3DKnock$gene_sign

}
