% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneScan3D.R
\name{GeneScan3DKnock}
\alias{GeneScan3DKnock}
\title{GeneScan3DKnock: Knockoff-enhanced gene-based test for causal gene discovery (knockoff filter).}
\usage{
GeneScan3DKnock(
  M = 5,
  p0 = GeneScan3DKnock.example$GeneScan3D.original,
  p_ko = cbind(GeneScan3DKnock.example$GeneScan3D.ko1,
    GeneScan3DKnock.example$GeneScan3D.ko2, GeneScan3DKnock.example$GeneScan3D.ko3,
    GeneScan3DKnock.example$GeneScan3D.ko4, GeneScan3DKnock.example$GeneScan3D.ko5),
  fdr = 0.1,
  gene_id = GeneScan3DKnock.example$gene.id
)
}
\arguments{
\item{M}{Number of multiple knockoffs.}

\item{p0}{A N-dimensional vector of the original GeneScan3D p-values, calculated using GeneScan3D.KnockoffGeneration() function.}

\item{p_ko}{A N*M matrix of M knockoff GeneScan3D p-values, calculated using GeneScan3D.KnockoffGeneration() function.}

\item{fdr}{The false discovery rate (FDR) threshold. The default is 0.1.}

\item{gene_id}{The genes id for N genes considered in the analysis. Usually we consider N=~20,000 protein-coding genes.}
}
\value{
\item{W}{The knockoff statistics for each gene.}

\item{Qvalue}{The q-values for each gene.}

\item{gene_sign}{Significant genes with q-values less then the fdr threshold.}
}
\description{
This function performs the knockoff filter, and computes the q-value for each gene. This function takes the results from the GeneScan3D.KnockoffGeneration() function and get knockoff statistics and q-values.
}
\examples{
library(GeneScan3DKnock)

# Load data example
data("GeneScan3DKnock.example")

result.GeneScan3DKnock=GeneScan3DKnock(M=5,
p0=GeneScan3DKnock.example$GeneScan3D.original,
p_ko=cbind(GeneScan3DKnock.example$GeneScan3D.ko1,
          GeneScan3DKnock.example$GeneScan3D.ko2,
          GeneScan3DKnock.example$GeneScan3D.ko3,
          GeneScan3DKnock.example$GeneScan3D.ko4,
          GeneScan3DKnock.example$GeneScan3D.ko5),fdr = 0.1,
          gene_id=GeneScan3DKnock.example$gene.id)
result.GeneScan3DKnock$W
result.GeneScan3DKnock$Qvalue
result.GeneScan3DKnock$gene_sign

}
